requirements.txt
fastapi[all]==0.94.1
mangum==0.17.0
# boto3==1.34.138
# botocore==1.34.138
exceptiongroup==1.2.1
shortuuid==1.0.13
pyjwt==2.8.0
pytz==2024.1